scraper/requirements.txt
crawl4ai>=0.8.0
tenacity
pyyaml
pytest
pytest-asyncio
